requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiokafka==0.8.0
aiopg==1.4.0
aiosignal==1.3.1
aiosmtplib==1.1.7
anyio==3.6.2
APScheduler==3.10.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
Babel==2.9.1
black==22.12.0
blinker==1.5
certifi==2022.12.7
charset-normalizer==2.1.1
defusedxml==0.7.1
Django==3.2.17
django-allauth==0.52.0
django-apscheduler==0.6.2
django-crispy-forms==1.14.0
django-filter==22.1
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.86.0
fastapi-mail==1.2.0
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
html2text==2020.1.16
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
kafka-python==2.0.2
magic-filter==1.0.9
multidict==6.0.4
mypy==0.991
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.2
pytest==7.2.0
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
sniffio==1.3.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
starlette==0.20.4
tomli==2.0.1
types-requests==2.28.11.8
types-urllib3==1.26.25.4
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uvicorn==0.19.0
yarl==1.8.2
